refactor: extract swap logic to new function
Refactored swap logic from `Pool` into a new function `v3_swap` in `swap_math.rs`. This improves modularity and makes the code more maintainable. All relevant references and usages have been updated accordingly.
