+++ src/entities/pool.rs
@@ -1,5 +1,5 @@
 use crate::prelude::{Error, *};
-use alloy_primitives::{ChainId, B256, I256, U160, U256};
+use alloy_primitives::{ChainId, B256, I256, U160};
 use core::fmt;
 use once_cell::sync::Lazy;
 use uniswap_sdk_core::prelude::*;
@@ -51,25 +51,6 @@ where
             && self.liquidity == other.liquidity
             && self.tick_current == other.tick_current
     }
-}
-
-struct SwapState<I = i32> {
-    amount_specified_remaining: I256,
-    amount_calculated: I256,
-    sqrt_price_x96: U160,
-    tick: I,
-    liquidity: u128,
-}
-
-#[derive(Clone, Copy, Default)]
-struct StepComputations<I = i32> {
-    sqrt_price_start_x96: U160,
-    tick_next: I,
-    initialized: bool,
-    sqrt_price_next_x96: U160,
-    amount_in: U256,
-    amount_out: U256,
-    fee_amount: U256,
 }
 
 impl Pool {
@@ -276,114 +257,17 @@ impl<TP: TickDataProvider> Pool<TP> {
         amount_specified: I256,
         sqrt_price_limit_x96: Option<U160>,
     ) -> Result<SwapState<TP::Index>, Error> {
-        let sqrt_price_limit_x96 = sqrt_price_limit_x96.unwrap_or_else(|| {
-            if zero_for_one {
-                MIN_SQRT_RATIO + ONE
-            } else {
-                MAX_SQRT_RATIO - ONE
-            }
-        });
-
-        if zero_for_one {
-            assert!(sqrt_price_limit_x96 > MIN_SQRT_RATIO, "RATIO_MIN");
-            assert!(sqrt_price_limit_x96 < self.sqrt_ratio_x96, "RATIO_CURRENT");
-        } else {
-            assert!(sqrt_price_limit_x96 < MAX_SQRT_RATIO, "RATIO_MAX");
-            assert!(sqrt_price_limit_x96 > self.sqrt_ratio_x96, "RATIO_CURRENT");
-        }
-
-        let exact_input = amount_specified >= I256::ZERO;
-
-        // keep track of swap state
-        let mut state = SwapState {
-            amount_specified_remaining: amount_specified,
-            amount_calculated: I256::ZERO,
-            sqrt_price_x96: self.sqrt_ratio_x96,
-            tick: self.tick_current,
-            liquidity: self.liquidity,
-        };
-
-        // start swap while loop
-        while !state.amount_specified_remaining.is_zero()
-            && state.sqrt_price_x96 != sqrt_price_limit_x96
-        {
-            let mut step = StepComputations {
-                sqrt_price_start_x96: state.sqrt_price_x96,
-                ..Default::default()
-            };
-
-            // because each iteration of the while loop rounds, we can't optimize this code
-            // (relative to the smart contract) by simply traversing to the next available tick, we
-            // instead need to exactly replicate
-            (step.tick_next, step.initialized) = self
-                .tick_data_provider
-                .next_initialized_tick_within_one_word(
-                    state.tick,
-                    zero_for_one,
-                    self.tick_spacing(),
-                )?;
-
-            step.tick_next = TP::Index::from_i24(step.tick_next.to_i24().clamp(MIN_TICK, MAX_TICK));
-            step.sqrt_price_next_x96 = get_sqrt_ratio_at_tick(step.tick_next.to_i24())?;
-
-            (
-                state.sqrt_price_x96,
-                step.amount_in,
-                step.amount_out,
-                step.fee_amount,
-            ) = compute_swap_step(
-                state.sqrt_price_x96,
-                if zero_for_one {
-                    step.sqrt_price_next_x96.max(sqrt_price_limit_x96)
-                } else {
-                    step.sqrt_price_next_x96.min(sqrt_price_limit_x96)
-                },
-                state.liquidity,
-                state.amount_specified_remaining,
-                self.fee as u32,
-            )?;
-
-            if exact_input {
-                state.amount_specified_remaining = I256::from_raw(
-                    state.amount_specified_remaining.into_raw() - step.amount_in - step.fee_amount,
-                );
-                state.amount_calculated =
-                    I256::from_raw(state.amount_calculated.into_raw() - step.amount_out);
-            } else {
-                state.amount_specified_remaining =
-                    I256::from_raw(state.amount_specified_remaining.into_raw() + step.amount_out);
-                state.amount_calculated = I256::from_raw(
-                    state.amount_calculated.into_raw() + step.amount_in + step.fee_amount,
-                );
-            }
-
-            if state.sqrt_price_x96 == step.sqrt_price_next_x96 {
-                // if the tick is initialized, run the tick transition
-                if step.initialized {
-                    let mut liquidity_net = self
-                        .tick_data_provider
-                        .get_tick(step.tick_next)?
-                        .liquidity_net;
-                    // if we're moving leftward, we interpret liquidityNet as the opposite sign
-                    // safe because liquidityNet cannot be type(int128).min
-                    if zero_for_one {
-                        liquidity_net = -liquidity_net;
-                    }
-                    state.liquidity = add_delta(state.liquidity, liquidity_net)?;
-                }
-                state.tick = if zero_for_one {
-                    step.tick_next - TP::Index::ONE
-                } else {
-                    step.tick_next
-                };
-            } else if state.sqrt_price_x96 != step.sqrt_price_start_x96 {
-                // recompute unless we're on a lower tick boundary (i.e. already transitioned
-                // ticks), and haven't moved
-                state.tick = TP::Index::from_i24(state.sqrt_price_x96.get_tick_at_sqrt_ratio()?);
-            }
-        }
-
-        Ok(state)
+        v3_swap(
+            self.fee.into(),
+            self.sqrt_ratio_x96,
+            self.tick_current,
+            self.liquidity,
+            self.tick_spacing(),
+            &self.tick_data_provider,
+            zero_for_one,
+            amount_specified,
+            sqrt_price_limit_x96,
+        )
     }
 }
